refactor: Changing requests namespace.
