+++ app/Http/Requests/User/Store.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\User;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class CooperativeAdminStore extends FormRequest
+class Store extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -25,10 +25,16 @@ class CooperativeAdminStore extends FormRequest
     {
         return [
             'name' => 'required|string',
-            'email' => 'required|email|unique:users,email',
+            'email' => 'required|email|unique:users',
+            'password' => 'required|string|min:6',
             'cpf' => 'required|regex:/^[0-9]{3}\.[0-9]{3}\.[0-9]{3}\-[0-9]{2}/|unique:users,cpf',
             'phones' => 'required|array',
             'phones.*.number' => 'required|string|regex:/^\([0-9]{2}\) [0-9]{5}\-[0-9]{4}/|distinct|unique:phones,number',
+            'addresses' => 'required|array',
+            'addresses.*.street' => 'required|string',
+            'addresses.*.neighborhood' => 'required|string',
+            'addresses.*.city' => 'required|string',
+            'addresses.*.number' => 'required|string'
         ];
     }
 }
